Search improvements - modules moved to "elasticsearch" subfolder + CHANGELOG update
