+++ core/lib/search/adapter/api/elasticsearch/mapping.js
