+++ core/lib/search/adapter/api/elasticsearch/score.js
